Avoid use of auxiliary class, and other warnings
* Move auxiliary AsciiSequence class to own file
* Remove redundant type args (use diamond)
* Suppress warnings in thrift classes
